webapps/requirements.txt
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-kms==2.5.0
APScheduler==3.5.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.0
blinker==1.4
brotlipy==0.7.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
constantly==15.1.0
crcmod==1.7
cryptography==2.3.1
cssselect==1.0.3
cx-Oracle==7.0.0
decorator==4.3.0
Django==2.1.1
django-apscheduler==0.2.13
django-cors-headers==2.4.0
django-mysql==2.4.1
djangorestframework==3.8.2
DtCrawlEngine==1.2.6
enum-compat==0.0.2
et-xmlfile==1.0.1
fake-useragent==0.1.11
h11==0.7.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4
Jinja2==2.10
jsonpath-rw==1.4.0
jwcrypto==0.5.0
kaitaistruct==0.8
ldap3==2.5.1
lxml==4.2.5
MarkupSafe==1.0
mitmproxy==4.0.4
mysql-connector-python==8.0.12
numpy==1.15.2
opencv-python==3.4.3.18
openpyxl==2.5.12
oss2==2.5.0
pandas==0.23.4
parsel==1.5.0
passlib==1.7.1
Pillow==5.2.0
pip==9.0.1
ply==3.11
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycryptodome==3.6.6
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.6.4
pymongo==3.7.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyperclip==1.6.4
pyquery==1.4.0
python-dateutil==2.7.3
python-jwt==3.2.3
python-scrapyd-api==2.1.2
pytz==2018.5
queuelib==1.5.0
redis==2.10.6
redlock-py==1.0.8
requests==2.19.1
ruamel.yaml==0.15.63
Scrapy==1.5.1
scrapy-fake-useragent==1.1.0
scrapy-redis==0.6.8
scrapy-splash==0.7.2
scrapyd==1.2.0
selenium==3.14.1
service-identity==17.0.0
setuptools==36.5.0
six==1.11.0
sockjs-tornado==1.0.2
sortedcontainers==2.0.4
tornado==4.3
Twisted==18.7.0
tzlocal==1.5.1
urllib3==1.23
urwid==2.0.1
uWSGI==2.0.17.1
virtualenv==15.1.0
w3lib==1.19.0
wsproto==0.11.0
xlrd==1.1.0
zope.interface==4.5.0